requirements.txt
streamlit==1.32.0
tensorflow
numpy==1.26.4
Pillow==10.2.0